requirements.txt
mysql-connector-python
boto3
paramiko
scp
flask
requests
